Added service interface and improved documentation
